requirements.txt
flask
opencv-python-headless
requests
psycopg2-binary
python-dotenv